Move specializeNorm / specialise
The functions that perform most of the work for transformations
like constantSpec were defined across two Util modules. This can
be simplified and the simplified version moved to the
Clash.Normalize.Transformations.Specialize module where it is easier
to find.
